pyproject.toml
[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[project]
name = "seq-len-balance"
version = "0.1.0"
requires-python = ">=3.11"
description = "Fast sequence-length balancing: KK partition, FFD, and BFD bin packing"
dependencies = [
    "mpi4py>=4.1.1",
    "pyzmq>=27.1.0",
]

[tool.maturin]
features = ["pyo3/extension-module"]

[dependency-groups]
benchmark = [
    "matplotlib>=3.10.8",
    "pytest>=9.0.2",
]
gpus = [
    "torch",
    "psutil",
]


[tool.uv]
dev-dependencies = [
    "matplotlib>=3.10.8",
    "maturin>=1.0,<2.0",
    "numpy>=2.4.2",
]
